3ds text rendering

+++ include/utilities/functions.hpp
@@ -21,7 +21,7 @@ namespace love
         return path.has_extension();
     }
 
-    inline long get_line(const std::string_view& string, int start)
+    inline size_t get_line(const std::string_view& string, int start)
     {
         if (string.empty())
             return std::string::npos;
